requirements.txt
aiogram==3.3.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
pillow==10.2.0
pytest==7.4.3
gspread==5.12.3
google-auth==2.28.0
